Clean up NRPE
BY having includes all over the shop we make things harder to
read and harder to maintain, so just have a single source.  This
removes the temptation to put checks in conditional code, thus
causing alerts when machines are taken out of production for
testing purposes as the checks disappear.  Good enough until this
approach is proven to be too naive