Prepare for cephgasm
Update the ceph module to take a hash of key value pairs to make
adding new parameters easier.  For now this is set to include

  ceph::params::global_params:
    osd_max_backfills: 1
    osd_recovery_max_active: 1

globally.  I'm not entirely sure which of these the rebalance is
going to manifest itself as, but hey ho you live and you learn.
Tomorrow's push to master will add

  ceph::params::global_params:
    osd_crush_chooseleaf_type: 2

which as you can see from the decompiled crush map will distribute
objects across chassis and not nodes

  # types
  type 0 osd
  type 1 host
  type 2 chassis

Which will be verified by checking the PG dump upon completion.