Kill logstash spam with fire
Previously we tried to be clever and install log definitions where they
existed.  However this means we can't remove the top level logshipper
from environments without removing the configurations from profiles.
As all log shippers basically ignore files that aren't present we now
just include the lot. If you remove the log shipper, you remove the
configuration in a single swipe *bam*.  This also has the knock on
effect of making the openstack profiles totally nicer removing a load
of if statements