Get Foreman's SSL setup fixed
So the initial X509 problem was that foreman.core.sal01.datacentred had been
changed to a CNAME, so obviously foreman0.core.sal01.datacentred didn't match
so that's been fixed.  Secondly when puppet connects via a load balancer to
submot reports foreman tries to validate the SSL certificates.  For future
proofing purposes I've added an SSL leg from the gateways to foreman* and
updated the configuration to trust gateway* certificates.  Obviously in true
foreman style it only cares about the distinguished names and not the
foreman.core.sal01.datacentred alternative name.  Also I noted BMC required
the fqdn I removed before, so I've added it back but with a sane default and
an override for core