Move BMC status checks into dc_nrpe
Removes implicit requirement that to use BMC provisioning you
need to install NRPE.  As an added bonus, this gets rid of
scope.lookupvar() calls which silently return '' if you move the
source.  Having the variables defined locally in the class that
consumes them solves this by raising a compilation error.  Also
updates R220 firmware to 2.0f across the board.