Improve Postfix Monitoring
So we have an MX that accepts mail, queues it, and relays it slowly.  We have
a check that alerts that mail is being queued and relayed slowly: in essence we
are getting woken up at 4am for something that is working as designed.  For
now lets get rid of the stupid alert.  What Matt will rage about is that we are
not alerting when mail cannot be sent.  That information can be parsed from syslog
by riemann and alerted from there.  Regardless this is not customer impacting so
do not alert!