Switch to PPA log-courier
Uses the PPA log-courier package and init scripts.  Configuration is now
derived from YAML and just spat out buy the JSON formatter.  The reason
for this is that upstream log-courier will complain if we terminate the
list of files with an empty hash (e.g. the only way concat based config
can work).