Vagrant Improvements
Rename puppetdb to puppet, install the Rack puppetmaster package to create a
CA, allow autosigning for all hosts in vagrant.dev.  The client is updated to
point at the new hostname, have a valid puppet.conf and also run --test
against the master to generate SSL certificates.  The puppet DB instance also
uses the CA certificates on the master so we can do away with nginx proxy
hacks as there is trust from a single source.  Additionally &> doesn't seem
to work so I've reverted to > 2>&1 to be rid of red text