Replace collectd with telegraf in Core Services
This change swaps out collectd for an InfluxDB-native metrics collector
called telegraf.  Aside from performance benefits (as we'll be able to
disable the Graphite-compatibility plugin in InfluxDB), it offers
greater flexibility for how measurements are sent and allows additional
tags to be specified for each metric.  In our case this is three Facts -
domain, role, and hostgroup.

Stunnel is still required as telegraf doesn't (yet) support SSL
natively.