Fix /etc/login.defs so UID and GID ranges for system accounts are correct
The 'telegraf' package was installed with 'useradd', which doesn't have
values specified for minimum / maximum UID and GID.  This meant that it
ended up with a UID of 1006 and a GID of 1001 which is conflicting with
the range we've allocated for our user and group accounts.

This change also purges the 'collectd' and 'telegraf' packages from all
machines in Core.  Once the user account for the latter has been
removed, a second PR will be raised to re-install the package - this
time with a UID < 1000.