Install telegraf from DataCentred repo with fixed postinst script
The package supplied upstream has an error in the postinst script [0]
which means that the user still isn't created as a 'system' account, i.e
with a UID / GID < 1000.  I've fixed the script and recreated the
package which is now in our own repo.

This change includes the DataCentred repo for Trusty and amends the
configuration data so that telegraf is installed from here instead.

[0] https://github.com/influxdb/telegraf/issues/231