Parameterise common NRPE checks
Thresholds for certain checks need consideration on a per-role basis,
including OpenStack Network Nodes which have a large number of processes
that reflect the number of L2 and L3 namespaces created by customers.

This change allows a few of these values to be overridden - and in the
case of this PR values for `check_procs` is increased for the
`openstack_network` role.