Addition of module to configure knot dns (1.4.2) which is currently used solely on NS1.
Authoritive DNS, supports reverse DNS lookups too.
Currently configured manually for a single client, so automation is useful incase such a usecase occurs again.
Currently, it is necessary for the zone files to be added manually (/var/lib/zones) and this could be improved by fetching the files, or creating the zones by templates.
Semantic checks are done on all zone files to promote integrity.
Module for puppet v2 wasn't written at the time of commit, but might add security/features so could be something to do in the future.