Manage L3 agent configuration and enable automatic failover
For one reason or another we're not actually managing the configuration
for L3 agents, so we're basically running with defaults.  Thankfully
these are fairly sane, but for changes to this component defined in
cfa3cf3dcf1d2546edc7821c1da3587cfae89de4 to be recognised then this
class needs to be included in the profile.

This commit also adds the 'allow_automatic_l3agent_failover' option,
which enables us to quickly migrate all existing L3 agents from a downed
network node.

Also addresses PD-1748.