Improve Site Manifest
hiera_include is great at selectively adding classes to be included on a
host based on architecture, OS etc. but when it comes to exclusing classes
on a specific role it does not cope too well (e.g. you'd need to include
the class on every role bar one: not scalable, prone to coders losing class
inclusions when creating new roles).  What is better is specifying a class
for inclusion globally but having some mechanism to exculde it for a specific
class of machine or role.  In this case we want monitoring and mail on all
the things except the core gateway which proxies on the required ports. This
change simply loads classes with a hiera_array call, loads exclusions with
another hiera_array call, then includes the boolean difference: sexy.