Remove Package Defaults From DC Icinga
Some packages installed by dc_icinga clash with dc_icinga2.  Even though both
are installed with ensure_packages() puppet sees these as definitions of
different resources because one declaration has a default parameter which
ensure the package is installed.  As this is the default action anyway just
remove the defaults so that the resources don't clash