Increase Config Timeout on Icinga2
At present there are a few thousand resources flying about from puppetdb and
compilation is slow, in the order of 10s of minutes.  This triages the problem
by increaing the client timeout.  The longer term fix is probably to apply
service checks to exported hosts based on a role host variable, but this needs
testing in my test environment.