Update the Configuration Mode of Icinga2
A bottom up configuration resulted in (N x machines x checks) resources
floating about which causes icinga to take 600s to run.  This isn't
scalable so I've moved to a top down model.  Hosts export their host record
only, global templates are defined on the master, these are applied to hosts
that match various critera (e.g. role) and are executed on the host zone.
Host checks are performed as before by the parent zone.  Configuration is
propagated from the master to each client pretty much instantly once it is live
on the master node.