Fix Foreman's setting of 'next-server' when creating DHCP entries
When Foreman creates a DHCP lease, it does so with an extra option to
override the 'next-server' (i.e the TFTP server) setting.  Unfortunately
it sets this to the MAC address of the Smart Proxy itself, which in our
case is actually a VIP.  This means that TFTP times out and the machine
doesn't provision.

This change removes the bit of code which adds this option, the
assumption being that whatever is set in the DHCP configuration itself
for a given subnet is the right next-server from which we should boot.