Update Icinga2 Base Checks
Jenkins returns a 403 and fails when checking / so check against /login instead.
Puppet at the moment doesn't allow access to the /status API so revert to a plain
TCP check for now until this is enabled, or we switch to puppet server. Finally
make load average checks relative to the number of CPUs installed in the
machine