[Icinga2] Get Ceph Mon and OSD Checks Working
Now each host exports a hash of interfaces we can apply these checks only
to the interfaces which are on the ark-ceph-public subnet.  For both of
these checks the ceph public IP address is extracted from the interface
attributes to perform the check against.  Additionally for the monitor
check we need to specify the monitor ID which is extracted from the
host.name variable using built in string manipulation