[Icinga] Restore Sanity to PSU Checks
The X9 class of SuperMicro hardware now exposes PSU status via ipmitool.
This adds in a new check to look for all PSUs on a system and report on
them.  Adds a new hostgroup for X9 machines and applies the check to all
new OSD nodes and existing compute nodes.