[Icinga2] Fix Satellite-Agent Configuration
The current model relies on all hosts within a domain to report into an
Icinga satellite or master operating in the same domain.  Young Tristan
has introduced a third class whereby a host from one domain tries to
connect to a satellite or master in a different domain, which broke all
the things.  Allow agents to optionally connect to a master or satellite
in a different domain by allowing the DNS driven endpoint host address
to be overridden to point at a public address.