Make Vagrant 1.8 a requirement, add new options
This commit makes Vagrant 1.8 the minimum and enables support for linked
clones which should speed up provisioning.  Note that you have to
destroy / create instances for this option to come into effect.

It also removes the options for port forwarding, this is largely
redundant given that every guest has bridged networking configured and
is wholly accessible from the host OS.