[Icinga2] ATA SMART Checks
Installs the check_scsi_smart command from our package repo, defines a
check command wrapper for this binary and applies the service to all
block devices defined on the host.  ARM nodes are explicitly blacklisted
as SMART commands slay the system.  Admittedly this was based on the
old ioctl code and not the SCSI generic stuff so it may be improved.