[Icinga2] Block Device Naming Fix
When the smart service is applied for each block device an object name of
'smartcciss!c0d0' is generated.  Unfortunately the '!' character is illegal
as an object name, so translate globally so the object name is now legally
'smartcciss/c0d0'.  The display name has also been updated to appear more
user friendly and the necessary template changes pushed upstream.  A more
general fix will happen when I get the time to architect, engineer and test