[Puppet] Add Small Splay To Puppet
The puppet module schedules runs at a minute granularity based on the
hashed MAC address.  This leads to very spiky performance characteristics
as N/30 machines all try to connect at the same time every minute.  Add
in a small 60 second splay to the agent configuration to see if this evens
out the CPU utilisation graphs