[Icinga2] Postgesql Initial Check
Rolls out the built in pgsql check against all postgres roles.  I've moved
the password in hiera so it can be shared by the global service definition
however there remains the fact that the password is encoded plain text in
dc_icinga.  This will be addressed when the existing code is sunset