[Icinga2] Mail Queue Checks
Add in a generic set of mail queue length checks for all nodes except
the gateways which cant run an MTA.  Check values are good enough for
both normal MTAs and smart hosts to detect a communications disruption
between * and mx0/1 and mx0/1 and google.

Also sneaks in a change to make the Postgres checks active on the host
for distributed alerting rather than leave the icinga servers as a SPOF
in a passive setup.