[Icinga2] Add Generic OS Plugin
Most of the OpenStack checks simply check that the process is running
and that it has an AMQP connection back to RabbitMQ.  This check does
the same (pgrep for the process and look for a destination port of
5672), but cuts down on the code duplication by parameterising the
process name to search for.  With a few tweaks it may work for all
openstack services rather than just doing a basic process check e.g.
conductor doesn't work as the parent process is the first to match
whereas the children have the AMQP connections.