[Ceph] Alias Storage Nodes
One of the 'new' nodes has the product name 'SSG-6027R-E1R12L' which is
a full product offering and not the existing board we have been using
('X9DRD-7LN4F(-JBOD)/X9DRD-EF').  Add a symbolic link in to say they
are the same kit without duplicating configuration.