[Icinga2] SAS Phy Checks
Applies a check to all nodes which searches for SAS phys on the system
and interrogates the error counters.  The counters cannot be reset so
we check for increases based on historical data collected at the last
invokation of the service check.  This class of warnings is indicative
of transport errors e.g. faulty cables or poorly seated drives