[Icinga2] Neutron Server Checks
Renamed check_nova_service to check_openstack_service as that functionality
was also present in the existing neutron server check and can be reused
to validate the process is up and connected to AMQP.  The API checks are
now just a basic keystone token get and neutron network request.  The check
also takes performance metrics so this can be injected into InfluxDB and
trends identified.  Additionally it has configurable warning and critical
thresholds to alert that doom is impending.