[BMC] Make User Provisioning Sane
This is currently failing on all new X9 storage servers.  Half of the
problem is having to hard code LAN channels and user IDs and doing
different things on different platforms.  This makes the problem go away
by interrogating ipmitool and finding a free slot to allocate the user
into.  As this is done via a type and provider we can actually change
permissions and passwords with puppet, even purge existing users...