[Icinga] Add A Memory Check
Seemingly we don't actually have any memory alerts (lol) so
roll out the stock memory check which will kick in at 10% remaining.
This shouldn't be a problem on ceph as the disk cache is counted as
free (e.g. what uses the memory) and on compute nodes with a 90%
memory limit we expect people to over provision anyway and have a lot
of physical pages paged out.  All boxes report as OK when the check
is performed via ansible