[BMC] Support HP iLO Properly
As with all tier 1 vendors it's my way or the highway, and as such doing
anything with the ipmitool provider proves nonsensical.  Cue my pants being
lowered as Meg tightens her strap on.  Any how, I've added a new provider
which works with hponcfg when the bios vendor is HP.  I've added a few fixes
to the ipmi provider which improve debug potential.  Each provider now defines
a set of features they support (e.g. ilo or ipmitool) which is then by used
by the ipmi_user type to decide which properties are relevant depending on
the provider types prefetched on the system.  Good news is the type can now
be configured to support all possible platforms without having to add in any
HP specific control logic.