[DNS] Fix LDAP DNS Entry
The existing DNS entry is an A record, but this is tring to install a CNAME
so fix the hiera configuration not to keep trying to apply the change.
Now I consider it the gateways should have the correct DNS alternative names
to work with CNAMEs, but this is a larger change with more scope for breakage.