[ARM64] Working Configuration for Vivid/Systemd/Kilo
Add Configuration:
Adds in apt configuration for Ubuntu 15.04.  Packages have been compiled and
added to apt.datacentred.services (icinga2, check_scsi_smart).

Disable Crash Dumps:
These packages are only available for i386 and amd64.  Consult
http://packages.ubuntu.com/vivid/linux-crashdump for future reference.

Update Service Provider:
Vivid onwards starts to use systemd as the default service manager however
as puppetlabs only support LTS releases officially we need to coax the
ARM nodes to use systemctl to manage services.

Fix Nova Service:
Was hard coded to upstart so I've added the ability to allow this to
be overridden to systemd.

Upgrade OVS Module:
Allows us to override the installation of the DKMS module which
is not available for the 3.19 kernel in use on Vivid.