Change Cinder's default Availability Zone to match Nova
This commit updates the configuration options for Cinder so that the
default availability zone matches that for Nova ('Production').  This is
currently causing problems when attempting to boot an instance from an
image creating a new volume, as Nova's AZ is passed to Cinder and as
that AZ doesn't exist from the latter's POV the request fails.