Renable Icinga2
Reverts 9a49d0daf62568c490db0244302916a36e09ad24 which allows icinga2 to
start up again and be managed.  I'll upgrade all nodes from wherever they
are to the latest and greatest 2.4.4 via ansible.  Alan reported the OOM
conditions as occurring on 2.4.1, but I'll add in an explicit memory self
check to warn when things appear to be too aggressive with memory use.