[Icinga2] Disk IO Queue Checks
Adds in a plugin to check the specified device's IO queue for excessive backlog,
creates an icinga2 check command and then applies the check to each block device
on each host.  Warning and critical thresholds are optionally specified as a
comma separated list as defined by the monitoring plugins guidelines; the first
value refers to spinning rust, the second to sold state.