[Icinga2] EDAC Memory Checks
Intends to kill off the silly OpenStack memory check by generically
checking every node (where supported) for correctable-errors detected
by the EDAC module.  These typically fire when a DIMM is about to die.
I'd love to see what it looks like when a DIMM is seated and busted,
is this detectable?