More Vagrant-friendly OpenStack provisioning fixes
This commit moves the Ceph-related configuration for Nova out into its
own class so that it can be easily excluded when building OpenStack in
virtual environments, i.e Vagrant.

It also adds a workaroud for the fact that Neutron L3 agent
configuration changes expect to notify a non-existent service;  It
should be notifying the VPN agent service instead.