[Icinga2] Sprinkling of Fixes
Corrected the performance data output on a couple of checks so that
the min and max values are displayed correctly in icingaweb2.  The
queue depth check was susceptible to the same 'HP are dicks' bug
that plagued the latency check.  The queue depth check has also been
made to accept lower and upper threshold limits based on percentages
which is measured against the queue depth set in sysfs, so it now
works across all devices and configurations without gross hacks.