Enable Neutron L3-HA
This commit introduces the configuration necessary to allow the creation
HA routers via Neutron.  It also includes a couple of patches that allow
migration from legacy routers to HA and vice-versa.

The default is to schedule a minimum of two for HA and a maximum of
three, which is in-line with our existing physical deployment.

Implements PD-2011.