[Icinga2] MD RAID Checks
Ports the existing check to python and enhances it slightly.  Now it will report
and alert separately on inactive array, failed devices or resynch and recovery
operations.  The synchronisation state of an array is also exported as performance
data, which may be helpful in correlating failures over time and predicting repair
rates via a pretty graph interface.