[Icinga2] MTU Checks
Distilation of the old check.  Applies the check to all interfaces with an
L3 address of 10.x.x.x, with the exception of virtual machines.  The check
looks at sysfs and ensures /sys/class/net/p1p1/mtu matches 9000.  Unlike
the previous check this one doesn't rely on netifaces which means it can be
pylinted - yay!

Fixes a bug with the active users check not returning the correct code too
and raising an exception when no sessions are detected.