[Icinga2] Add HAProxy Checks For Core
Installs a reimagining of the existing check that caters for mutual X.509
verfication.  This raises a warning if a member goes down and a critical
alert if a frontend or backend goes down.  Also accepts a comma separated
list of perfdata to export.  At present this is just the number of sessions.