[Icinga2] S3/OS HAProxy Checks
Updates the check script to accept a port if necessary and also perform some
basic error checking on the HTTP status code.  Icinga2 module updated to accept
a display name for 'apply Service' objects, so that regardless of the configuration
haproxy reports as haproxy.  Added in a new service check for port 1936 applied
locally to all S3 and OS public API endpoints.