[Icinga2] Check Cleanups
Updates check argument parsing to perform implicit type conversion and
also allow options that receive multiple parameters to be specified
multiple times and appended to a list.  Affected icinga2 checkcommands
have been updated to allow repeat keys when an array is specified as
the value, and also update the service checks to supply an array where
necessary e.g. user lists and read/write latencies for HDD/SSD.  The
icinga2 module is updated to support arrays in the configuration
formatter.