[Puppet] Stop Puppet Agent
Annoyingly Foreman (notice the pattern of destruction) uses puppet agent (the
service) to run the initial convergence upon reboot.  The puppet module expects
a blank slate and just the puppet-common package.  The long and short is we have
both cron and puppet agent running 4 times an hour, thus doubling our load.  This
stops the puppet agent.  Theoretically as we don't use MCollective any more we
could go back to using puppet agent, but that's a bigger job for this time on
the cusp of a bank holiday when the beers are calling.