[Icinga2] DHCP Checks Primary Interface
Updates the check to apply the check to all interfaces on the system.  The
check is only performed where the server has a DNS role and the interface
address is that of the hosts's primary address.  This removes the hard coded
reliance on bond0 and should now work on all DHCP servers.