[MMR1-OOB] Add MMR2 Subnet
MMR1 is a bit of a bad name as this is defining the entire OOB network across
the DWDM network.  This adds in the DNS range so we can specify names for
switches, routers and Foreman hosts, and a DHCP range for Forman to provision
in to.  IPAM updated to take this into account.