[Icinga2] Nova Service Check
Uses Keystone to authenticate and perform service discovery.  Proceeds
to spawn an instance and delete it reporting the times against warning
and critical thresholds, which is all exported via performance data so
it can be plotted over time via InfluxDB for predictive analysis. The
service checks against the API have been made to occur less frequently
to avoid bombarding the API and to cater for potentially longer check
times.  To give alerts in a timely manner I've reduced the limit beween
soft and hard alerts to 3 checks e.g. 15 minutes.