[Icinga2] Allow Satellites to Collect Other DNS Zones
In a sane world as I had pictured a satellite exists in a DNS zone and collects
zone configuration (zones/endpoints) exported by hosts in that DNS zone.  However
as the OOB network consists of mmr1.sal01, mmr2.sal01, xyz.mccp, xyz.mcob, xyz.mspk
and xyz.mspw we need the satellite that sits in mmr1 to be able to allow connections
from mmr2 etc.  The other option is to spin up a satellite in each POP, but for now
this will suffice until we determine it's shit and should be burned.