[Icinga2] Use Builtin RAID Check
Discovered that check_raid is available in nagios-plugins-contrib while
trying to fix a crash in the proprietary check.  This checks all mdX
arrays in the system, if any exist, and warn if out of sync or failed.
As such we should use this one as it's less of a maintenance overhead
and is less buggy, even though it doesn't support performance data, which
is of little utility anyway other than drawing pretty pie charts.