[Facter] Make $::mounts Sane
I noted that osnet* was constantly updating, apparently collectd was caring
about /run/netns/q* as a mount point.. omg.  Narrowing the source of the problem
down to some 3rd party code, it seems that namespaces are indeed mounts of type
'nsfs' which need to be blacklisted in this fact.