[Ceph] Update Network Configuration
Changed so that p1p1 -> public, p1p2 -> lb is now... em1 -> management,
p1p1 -> lb, p1p2 -> public, which changes $::ipaddress to be 10.10.112.x
as we'd expect, and be configured the same as every other ceph node.  As
a result the interface ordering for keepalived needs swapping, and also
the smell code needs deleting so it doesn't keep overwriting my interface
changes.