[Icinga2] RAID Check HP Fix
A shit tonne of our HP RAID controllers are warning about cache batteries
not being charged and thus the caching being disabled.  As a result the
output from cciss_vol_status is 1 thus causing a critical status *sigh*.
This makes the check less aggressive and just checks that the array is
OK, not the return code.