Enable the AggregateMultiTenancyIsolation Nova scheduler filter
This filter isolates tenants to specific host aggregates. If a host is
in an aggregate that has the filter_tenant_id metadata key, the host
creates instances from only that tenant or list of tenants. A host can
be in different aggregates. If a host does not belong to an aggregate
with the metadata key, the host can create instances from all tenants.

This allows us to drop a particular compute node into a Host Aggregate
away from Production so that it can be tested in isolation, i.e in the
event of some kind of maintenance.