Pin version of Ceilometer
The latest Ceilometer packages available in the Cloud Archive Kilo repo
introduce a bug when combined with the version of `oslo-utils` that's
we've also installed.  Details here:

https://bugs.launchpad.net/ceilometer/+bug/1556469

The fix hasn't yet made its way into the official packages yet, so for
now this commit ensures that the working version of Ceilometer is
installed in the relevant places.

As the package itself has been removed from upstream repos, I've made it
available in our own 'backports' repo on `apt.datacentred.services`.