Deprecate Neutron database consistency checks
This check was put in place to monitor for a problem we encountered a
couple of years back running Havana in which we saw duplicate
`segmentation_id`'s assigned to more than one network.

We haven't come across this problem since - the underlying bug has almost
certainly long since been addressed - and so this commit removes the
check so it's less legacy code for us to maintain.