Create concatenated client cert / key for use by haproxy (#2717)
The piece of code responsible for creating this file (which haproxy
needs for X509 auth) looks to have been accidentally deprecated as part
of #2573.  This commit simply replicates the missing bit of
functionality so that Puppet can at least run on haproxy loadbalancers
when being provisioned from scratch.

Helps with PD-2239.