[Quanta] BMC Support
Makes the ipmi_user ipmitool provider a bit more generic in that blank users
are reported as 'anonymous' in 'puppet resource ipmi_user'.  When a user is
made 'absent' this now just sets the permission to 'NO ACCESS' as it seems
a little more generic than trying to blank the user name, which doesn't work
on Quanta.  As such users with 'NO ACCESS' are filtered out when the resources
are pre-fetched.  Additionally when we are searching for a ID to create a user
as, there is now some additional logic to reuse a disabled user account if the
names match.