[Icinga] Exclude External IF From MTU Checks
Ceph backups were failing as the clients were sending jumbo frames over the public
internet.  I've manually set this to 1500 on the gateways and it appears to work
again.  However due to our old anal icinga checks this would cause an alert, which
is utterly pointless, so rectify this by excluding the public interface from the
checks.