[DNS] Update DNS Module (#2737)
There is too much of a disparity beteween this 3 year old code and what is on the
forge, which is at odds when trying to keep over-cloud provisioning in sync with
the cloud-cloud.  This updates to the latest upstream theforeman-dns module, thus
cutting down on proprietard clutter.  As a result a lot of the interfaces have
been improved significantly thus we benefit from cleaner code!  The two things we
miss out on is populating all NS records on provision, which will be handled by
a static record and an update to the dns_record type.  And the SOA records only
contain the host serving the request.  If nothing else this shows some much needed
love to archaic smell code.