[DNS] Fix MX Record exists? Check (#2751)
At the moment the provider is basically asking if an exchange exists that matches the
name e.g. if 'sal01.datacentred.co.uk' in ['mail.datacentred.services'], which doesn't
work and the record gets added each puppet run.  This updates the check to see whether
the rdata field exists within the array of returned exchanges.