[DNS] Refactor Provider (#2753)
Actually started out as a bug in the to_rdata code for MX records which incorrectly
returned the preference and not the exchange, affects all domains.  What actually
happened was the much needed refactor to bring things back into a state of sanity.
As such MX records are defined for the domain (e.g. the resource name) and not a
fake host.  Exchanges and preferences are passed in optionally as an arbitrarily
sized array.  Likewise SRV records get similar treatments.