Enable RedHat Enterprise Linux in Vagrant (#2768)
This commit switches testing of RHEL-based virtual machines away from
CentOS and onto pukka RedHat 7.2.

Two things to be aware of:

* The Vagrant box I've created is VMware Fusion only for now -
there'll be a Virtualbox version soon (for anyone that cares).  For
reference, it's based off Chef's [Bento](http://github.com/chef/bento)
Packer templates with a couple of fixes for interface naming.

* The a requirement to install
[this](https://github.com/projectatomic/adb-vagrant-registration)
Vagrant plugin if you want to provision RHEL guests.  You'll also have
to speak to myself or Matt if you want a RHN account in order to enable
guest registration.

It also updates the repos we use for RHEL and OpenStack to use our own
mirror of the official packages.