[Icinga2] Fix DHCP (#2767)
These broke in core due to bond0 changing to em1.  We were previously seeding the
interface with that on the remote machine, aka a hack :)  This derives the primary
interface from each machine and adds it to the host metadata.  The check is then
applied to each DHCP host, the interface is looked up based on the NodeName global
variable locally, so it works for all endpoints within the checking zone.