[Icinga2] Update OS Endpoint Service Host
Spotted that some openstack checks were reported as being for icinga2.core.  These
have been updated to be applied to the OS vhost, so that it's instantly apparent
that the source of the issue is openstack either via the hostname or the little
symbol.

Admittedly seeing as the checks are hard coded to compute.datacentred.io the checks
should be updated to consume their public fqdn from host metadata.  But I'll leave
this for another day.