Configure ntpd to only attempt binding to loopback interface
ntpd won't start on certain nodes with a proliferation of virtual
interfaces i.e compute nodes.  This is because of braindead ntp defaults
in which it attempts to bind to all available interfaces.

This commit updates the associated Puppet module and adds in the
configuration option to tell ntpd to ignore all interfaces and only
listen on `lo`.

Fixes PD-2410.