[Puppet4] Working DNS Master Environment
Upgrades concat to 2.2.0 which actually works with Puppet 4.  As a result pupmod-concat
had to be deleted due to a namespace clash (wasn't actually used) and DHCP needed to be
suitably fixed as it's several decades out of date.  Foreman no longer has control over
Puppet certificates, we manage this in a setup stage.  Actual puppet configuration is
selected via hiera, driven by the major version so 3 & 4 can cohabit.  Added the PC1 repo
to trusty (no risk in upgrades as the AIO package is puppet-agent and not puppet-common).
Radius was broken as regards pacakge installation, and was written in some arcane style
so I've made this sane.  Lsyncd needed its service resource updating to use the debian
provider rather than init.  Finally an NRPE check was being installed in the wrong
location causing fails, putting it in the right place means no other dependencies are
needed.