Staging Core Fixes (#2848)
Postgres Provisioning Fixes:
Allow excluded classes to be specified on the command line and aggregated with the
hiera excludes.  Things like icinga2 will not provision correctly without store
configs.  The hiera was invalid for postgresql slaves so this has been fixed to
change a Fixnum into a String.  The provisioning code was incomplete and generally
unsafe so I've made a few modifications to make this work more reliably.

Foreman Prerequsites:
Installs a required yaml configuration file which defines memcached configuration
for a single foreman instance.

Foreman Fixes:
Pinning the mod passenger binary causes provisioning to fail as it does not
automatically pull in the matching passenger package.  This removes the pin
and just uses the most recent version, no point in living in the past. The
upstream 1.11 release requires a binary hack in order to work with ruby
2.x.x as 1.9.x has been deprecated.