[Docker] Configure Docker Registry
This commit introduces basic configuration for Docker Registry, which will use
the Swift API in order to store the actual images itself.  For now it'll be
deployed on `graph` which currently hosts Grafana and nothing else (and so has
the `telemetry` role).

A new project, `devops`, and a new storage container, `registry`, has been
created for this use.

There'll be a subsequent commit to introduce the necessary haproxy
configuration which will enable SSL cert-based auth for Docker clients.

Part of PD-1962.