[BMC] Platform Agnostic Orchestration (#2877)
The old python based approach doesn't work on RedHat and certainly will not work
across versions of all our operating systems due to the library being installed
in a platform dependant location (i.e. /usr/lib/python2.7).  This kills that in
the face and moves this functionality into a custom type.  A number of improvements
as a result, it will create the interface if it doesn't exist in foreman, much
better error checking, and it makes BMC username and passwords properties that
can be changed via the type.