Conditionally insert the X-Forwarded-Proto header for https
Update HAProxy's configuration on the OpenStack loadbalancer nodes to
pass the X-Forwarded-Proto header denoting https for all SSL
interactions.

This fixes the previously broken behaviour in OpenStack in which
Keystone had to be configured to return the full URL for the service.
This has caused problems in the past which let to the split-horizon
workaround. Further details:

https://bugs.launchpad.net/keystone/+bug/1370022
https://bugs.launchpad.net/keystone/+bug/1528258kkk

It won't have any effect right now, but will be necessary once we've
moved to a Mitaka-based Keystone Docker container.