[Puppet4] Rewrite dc_tftp
Horror code from back in the day... This beautiful module used sshkeys to generate
an ssh key, export it via a fact and query puppetdb on the other end to get a
public key.  All of which epically fails on Puppet 4.  This replaces dynamic keys
with static ones defined in hiera.  Also took the liberty to rewrite the code so
it's slightly more modern, standards compliant, cleaner and safe.  Updates
duplicity too as there was some crap code in there which was passing file modes
about as an integer :/