[OpenStack] Return IP of primary interface according to Foreman
The `::ipaddress` fact can't be relied on as it's just the IP address of
the first network interface when they're ordered alphabetically.  This
results in breakage when things like `docker0` bridges are introduced as
the wrong IP address will be exported for collection on the OpenStack
loadbalancers.

This commit adds a custom function which returns the IP address of the
primary interface according to Foreman and so should solve this problem.