[OpenStack] Workaround wonky exported resources behaviour
Slightly amend the approach taken in 367749d and assign the value
returned from the `foreman_primary_ipaddress` custom function to
variable, and then use that as the value passed to the `ipaddresses`
parameter as part of the exported resources.

This is required because the value returned from that function is actually
evaluated on the node where this virtual resource is realised, resulting
in the wrong IP address being applied to each of the balancer member
stanzas.

Fixes PD-2516.